revert(servicenow): drop unsupported Xanadu+ FQCN redesign (#1062)
ServiceNow exposes no way to configure a custom fully-qualified class name for a
customer JAR-import external credential resolver (verified on an Australia instance:
no fqcn field, no resolver-class field, no custom vault/resolver definition table).
The MID Server always loads a customer resolver as the shared
com.snc.discovery.CredentialResolver, so an fqcn-only JAR would fail to load on Xanadu+.

Collapse the resolver back to the single com.snc.discovery.CredentialResolver class and
one JAR per release (removing the keeper FQCN class and the resolverVariant build/workflow
split). Keeps the KSM SDK 17.3.0 pin (fixes the pamSettings deserialization break) and the
pamSettings regression test, and preserves the release-branch CI work (SBOM, CDN-sourced
MID libs, per-release toolchain). Docs updated: register by the shared class name; only one
custom resolver JAR per MID Server.

Refs: IMP-3033, KSM-610
